// frontend/lib/api/generated/model/pedidoSemanalOut.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SIGA-UNMSM API
 * API del Sistema Integral de Gestión de Almacén y Abastecimiento Alimentario — UNMSM. Módulos activos: Autenticación, Usuarios, Almacenes, Catálogo Nutricional.
 * OpenAPI spec version: 0.1.0
 */

export interface PedidoSemanalOut {
  pedido_semanal_id: number;
  orden_compra_id: number;
  menu_id: number;
  almacen_id: number;
  almacen_codigo: string;
  almacen_nombre: string;
  semana_inicio: string;
  semana_fin: string;
  estado: string;
}